feat: move lsHandler out to its own module
